build progress

+++ server/src/main/java/edu/unh/cs/cs619/bulletzone/datalayer/account/BankAccount.java
@@ -8,7 +8,7 @@ public class BankAccount extends OwnableEntity {
 
     public double getBalance() { return balance; }
 
-    public BankAccount(int userId) {
+    public BankAccount(long userId) {
         super(new BankAccountRecord());
         this.balance = 1000.0; // Initial balance of 1000 credits
     }
